Fix bug in Alg1 that led to little exploration
In SimOOS there was this mechanism for exploring arms to fill the
N_t_aso if it was 0. The way I wrote code for Algorithm1 did not
have this part and just used N_t_aso = min(1, its_value). This
resulted in not enough exploration even when costs are known.